[all] Google tests now again all OK

+++ src/lcd.cpp
@@ -288,12 +288,12 @@ void lcd_outdezNAtt(xcoord_t x, uint8_t y, int16_t val, LcdFlags flags, uint8_t 
   x -= fw + 1;
 
   for (uint8_t i=1; i<=len; i++) {
-    div_t qr = div(val, 10);
+    div_t qr = div((uint16_t)val, 10);
     char c = qr.rem + '0';
     uint8_t f = flags;
     if (dblsize) {
       if (c=='1' && i==len && xn>x+10) { x+=2; f|=CONDENSED; }
-      if (val >= 1000) { x+=FWNUM; f&=~DBLSIZE; }
+      if ((uint16_t)val >= 1000) { x+=FWNUM; f&=~DBLSIZE; }
     }
     lcd_putcAtt(x, y, c, f);
     if (mode==i) {
@@ -315,7 +315,7 @@ void lcd_outdezNAtt(xcoord_t x, uint8_t y, int16_t val, LcdFlags flags, uint8_t 
         x--;
       }
     }
-    if (dblsize && val >= 1000 && val < 10000) x-=2;
+    if (dblsize && (uint16_t)val >= 1000 && (uint16_t)val < 10000) x-=2;
     val = qr.quot;
     x-=fw;
   }
